changes enabling 2D subsymm templating
First additions to the codebase for templating subsymmetric templates
into supersymmetries. Worked for reconstructing toy C2 into C2 problem
where the entire set of residues was within the template.

Did not work for first experiment in which first 100 residues from each
chain were "visible" in the template, and the rest did not contain
template information.